add a mode to `ElimIndPreds` that relies on pattern-matching
For smbc, selectors are to be banned because the conditions
might be evaluated in any order, and recursive definitions will
be non terminating.

For instance,

`pred[wf] even : nat -> prop where
  even 0;
  forall n. even n => even (S (S n)).`

will yield:

`rec even : nat -> prop where
  forall n. even n =
    n = 0 ||
    is-S n && is-S (pred n) && even (pred (pred n)).`

which is non-terminating in smbc because `&&` is orderless.
Instead, we compile into something roughly equivalent to

`rec even : nat -> prop where
  forall n. even n =
    (match n with 0 -> true | _ -> false) ||
    (match n with S (S x) -> even x | _ -> false).`